Fix the scheduler timeout logic
We don't need to loop, because the scheduler is the only goroutine sending on the channel.
The only thing the newly started goroutine is allowed to do is wait for us to send it work, so there's no point in handling the case where it's busy.

Additionally, have the timer prime its channel with a value when we actually get a value from it, so the next iteration's Stop will actually see the value on the C.

Tested using the test bot.